@@ -12,12 +12,12 @@
         }
     ],
     "require": {
-        "php": ">=5.5",
-        "zendframework/zend-mail": ">= 2.7.2",
-        "guzzlehttp/guzzle": "^6.0"
+        "php": "^8.1",
+        "guzzlehttp/guzzle": "^6.0",
+        "laminas/laminas-mail": "^2.23"
     },
     "require-dev": {
-        "phpunit/phpunit": "~3.7",
+        "phpunit/phpunit": "^10.0",
         "satooshi/php-coveralls": "^1.0"
     },
     "autoload": {
