@@ -9,11 +9,11 @@ namespace Checkdomain\Comodo;
  * @method int                         count()
  * @method int                         countMessages($flags = null)
  * @method void                        createFolder(string $name, $parentFolder = null)
- * @method \Zend\Mail\Storage\Message  current()
+ * @method \Laminas\Mail\Storage\Message  current()
  * @method array                       getCapabilities()
- * @method \Zend\Mail\Storage\Folder   getCurrentFolder()
- * @method \Zend\Mail\Storage\Folder   getFolders($rootFolder = null)
- * @method \Zend\Mail\Storage\Message  getMessage(int $id)
+ * @method \Laminas\Mail\Storage\Folder   getCurrentFolder()
+ * @method \Laminas\Mail\Storage\Folder   getFolders($rootFolder = null)
+ * @method \Laminas\Mail\Storage\Message  getMessage(int $id)
  * @method int                         getNumberByUniqueId(string $id)
  * @method array|string                getRawContent(int $id, $part = null)
  * @method array|string                getRawHeader(int $id, $part = null, int $topLines = 0)
